clang-tidy fixes

+++ tasks/rozenberg_a_bubble_odd_even_sort/mpi/src/ops_mpi.cpp
@@ -11,7 +11,7 @@
 namespace rozenberg_a_bubble_odd_even_sort {
 
 void RozenbergABubbleOddEvenSortMPI::LocalBubbleSort(InType &local_buf) {
-  int chunk = local_buf.size();
+  int chunk = static_cast<int>(local_buf.size());
   for (int i = 0; i < chunk; i++) {
     for (int j = 0; j < chunk - 1; j++) {
       if (local_buf[j] > local_buf[j + 1]) {
@@ -29,7 +29,8 @@ void RozenbergABubbleOddEvenSortMPI::ExchangeAndMerge(InType &local_buf, int nei
   MPI_Sendrecv(local_buf.data(), chunk, MPI_INT, neighbor, 0, neighbor_data.data(), neighbor_n, MPI_INT, neighbor, 0,
                MPI_COMM_WORLD, MPI_STATUS_IGNORE);
 
-  std::merge(local_buf.begin(), local_buf.end(), neighbor_data.begin(), neighbor_data.end(), merged.begin());
+  std::merge(local_buf.begin(), local_buf.end(), neighbor_data.begin(), neighbor_data.end(),
+             merged.begin());  // NOLINT(modernize-use-ranges)
 
   if (rank < neighbor) {
     std::copy(merged.begin(), merged.begin() + chunk, local_buf.begin());
